backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.51.0
pydantic==2.9.2
pydantic-settings==2.4.0
cryptography==43.0.1
loguru==0.7.2
httpx==0.27.0
groq==0.15.0
